requirements.txt
pytest
requests
pydantic
allure-pytest
jsonschema
faker
python-dotenv